update scripts for amplicon suite agggregator
