+++ src/AmpliconSuiteAggregator.py
@@ -1,371 +1,8 @@
-###
-# Author: Edwin Huang
-# This script takes in completed AmpliconArchitect and
-# AmpliconClassifier results and aggregates the results.
-###
-import argparse
-import sys
-import tarfile
-import os
-import re
-import pandas as pd
-import tarfile
-import shutil
-import json
-import subprocess
-import ast
-import zipfile
-
-
-
-class Aggregator():
-    def __init__(self, filelist, root, output_name):
-        self.filelist_fp = filelist
-        self.root = root
-        self.output_name = output_name
-        print(self.get_zip_paths())
-        self.unzip()
-        self.aggregate_tables()
-        self.json_modifications()
-        self.move_files()
-        self.cleanup()
-
-
-
-    def get_zip_paths(self):
-        """
-        Gets the individual file paths from the list of zip filepaths
-
-        calls from self
-        returns --> a list of filepaths to zip files
-        """
-        files = []
-        with open(self.filelist_fp, 'r') as filelist:
-            for line in filelist:
-                parsed = line.strip()
-                files.append(parsed)
-        filelist.close()
-        self.zip_paths = files
-        return files
-
-
-    def unzip(self):
-        """
-        Unzips the zip files, and get directories for files within
-        creates a dictionary that contains information to run make results table .py from Jens
-        deletes unnecessary files
-
-        returns -- None
-        """
-
-        # os.mkdir('all_AA_results')
-
-        for zip_fp in self.zip_paths:
-            fp = os.path.join(self.root, zip_fp)
-            try:
-                if ".tar.gz" in zip_fp:
-                    sample_name = os.path.basename(zip_fp).replace(".tar.gz", "")
-                    destination = f'./AA_outputs/{sample_name}'
-                    print(sample_name)
-                    with tarfile.open(fp, 'r') as output_zip:
-                        output_zip.extractall(destination)
-                    output_zip.close()
-
-                elif ".zip" in zip_fp:
-                    sample_name = os.path.basename(zip_fp).replace(".zip", "")
-                    destination = f'./AA_outputs/{sample_name}'
-                    with zipfile.ZipFile(fp, 'r') as zip_ref:
-                        zip_ref.extractall(destination)
-                    zip_ref.close()
-
-                ## Check if extracted contents is a directory,
-                ## and if there is only one directory. Make that the new sample_name
-                ## and move it upwards.
-                if len(os.listdir(destination)) == 1:
-                    for file in os.listdir(destination):
-
-                        if os.path.isdir(f"/opt/genepatt/AA_outputs/{sample_name}/{file}"):
-                            os.system(f"mv -v /opt/genepatt/AA_outputs/{sample_name}/{file} /opt/genepatt/AA_outputs/{file}")
-                            os.system(f"rm -r /opt/genepatt/AA_outputs/{sample_name}/")
-            except Exception as e:
-                print("The zip: " + fp + " ran into an error when unzipping.")
-                print(e)
-
-        samples = []
-        aa_result_paths = []
-        ac_input_paths = []
-        profiles_tsv_locations = []
-        cnv_calls_locations = []
-
-        make_results_dict = {}
-        # parse contents of zip to get sample_name
-        ## remove everything other than AA_results
-        for root, dirs, files in os.walk("./AA_outputs", topdown=False):
-            sample_name = ""
-            for name in dirs:
-                dir_name = os.path.join(root, name)
-                if "_AA_results" in dir_name:
-                    dir_name = dir_name.split("/")[-1]
-                    sample_name = dir_name.replace("_AA_results", "")
-                    sample_name = sample_name.replace('AA_outputs/', "")
-                    samples.append(sample_name)
-
-                    make_results_dict[f'sample_{sample_name}'] = {
-                    'name':sample_name,
-                    'ac_input':'',
-                    'profiles_tsv':'',
-                    'cnv_calls':''
-                    }
-
-
-                    aa_results_dir = os.path.join(root, name)
-                    aa_result_paths.append(aa_results_dir)
-                    os.system(f'bash /opt/genepatt/make_input.sh {os.path.join(root, name)} {os.path.join(root, f"{sample_name}_AC_inputs")}')
-                    make_results_dict[f'sample_{sample_name}']['ac_input'] = os.path.join(root, f"{sample_name}_AC_inputs.input")
-
-            for file in files:
-                if 'amplicon_classification_profiles.tsv' in file:
-                    profiles_tsv_location = os.path.join(root, file)
-                    profiles_tsv_locations.append(profiles_tsv_location)
-                elif 'CNV_CALLS.bed' in file:
-                    cnv_calls = os.path.join(root, file)
-                    cnv_calls_locations.append(cnv_calls)
-                elif ('.stderr' in file) or ('_run_AA_ac.sb' in file) or ("_run_PAA.sb" in file) or ("stderr" in file) or ('stdout' in file) or (".sb" in file):
-                    os.remove(os.path.join(root, file))
-
-
-
-
-        for sample in make_results_dict.keys():
-            name = make_results_dict[sample]['name']
-            for tsv in profiles_tsv_locations:
-                if name in tsv:
-                    make_results_dict[sample]['profiles_tsv'] = tsv
-            for cnv in cnv_calls_locations:
-                if name in cnv:
-                    make_results_dict[sample]['cnv_calls'] = cnv
-
-
-        # print(make_results_dict)
-        # run_make_results_table(make_results_dict)
-
-
-
-        # # move AA results to another directory and run results table there
-        # destination = os.path.join("/opt/genepatt/temp")
-        # if not os.path.exists("/opt/genepatt/temp"):
-        #     os.mkdir(destination)
-        # for aa_result in range(len(aa_result_paths)):
-        #     sample_dest = os.path.join(destination, samples[aa_result], f"{samples[aa_result]}_AA_results")
-        #     shutil.move(aa_result_paths[aa_result], sample_dest)
-        #     os.system(f'bash /opt/genepatt/make_input.sh {sample_dest} {os.path.join(destination, samples[aa_result], f"AC_inputs.txt")}')
-        #     os.system(f'python3 /opt/genepatt/')
-
-    def aggregate_tables(self):
-        """
-        From the unzipped paths, start aggregating results
-
-        """
-
-        output_head = ["Sample name", "AA amplicon number", "Feature ID", "Classification", "Location", "Oncogenes",
-                           "Complexity score",
-                           "Captured interval length", "Feature median copy number", "Feature maximum copy number",
-                           "Reference version", "Feature BED file", "CNV BED file", "AA PNG file", "AA PDF file",
-                           "Run metadata JSON"]
-        output_table_lines = [output_head, ]
-        aggregate = pd.DataFrame(columns = output_head)
-        cycle_files = []
-        graph_files = []
-        runs = {}
-        sample_num = 1
-        if not os.path.exists('results'):
-            os.mkdir('results')
-            os.chmod('results', 0o777)
-
-        # aggregate results
-        for root, dirs, files in os.walk("AA_outputs", topdown = False):
-            for name in files:
-                if "_result_table.tsv" in name:
-                    result_table_fp = os.path.join(root, name)
-                    try:
-                        df = pd.read_csv(result_table_fp, delimiter = '\t')
-                        aggregate = pd.concat([aggregate, df], ignore_index = True)
-                        runs[f'sample_{sample_num}'] = json.loads(df.to_json(orient = 'records'))
-                        sample_num += 1
-                    except:
-                        continue
-
-
-                    ## convert to json
-
-
-            for dir in dirs:
-                if "files" == dir:
-                    dir_fp = os.path.join(root, dir)
-                    os.system(f'cp {dir_fp}/* /opt/genepatt/files')
-
-
-
-        ## output the table
-        aggregate.to_csv('aggregated_results.csv')
-        aggregate.to_html('aggregated_results.html')
-        with open('results/run.json', 'w') as run_file:
-            json.dump({'runs': runs}, run_file)
-        run_file.close()
-
-
-    def aggregate_json(self, json_files):
-        """
-        Aggregates json files from separate samples
-
-        """
-        out_dict = {'run' : {}}
-        sample_ct = 1
-        for fp in json_files:
-            sample_dict = json.load(open(fp))
-            out_dict['run'][f'sample_{sample_ct}'] = sample_dict
-            sample_ct += 1
-
-        with open('run.json', 'w') as out_json:
-            json.dump(out_dict, out_json)
-        out_json.close()
-
-    def move_files(self):
-        """
-        Move files to correct location for output
-        """
-
-        # shutil.move('/opt/genepatt/files', '/opt/genepatt/extracted/files')
-
-        shutil.move('aggregated_results.csv', 'results/aggregated_results.csv')
-        shutil.move('aggregated_results.html', 'results/aggregated_results.html')
-        shutil.move('AA_outputs', 'results')
-
-
-    def tardir(self, path, tar_name):
-        """
-        compresses the path to a tar_mname
-        """
-        with tarfile.open(tar_name, "w:gz") as tar_handle:
-            for root, dirs, files in os.walk(path):
-                for file in files:
-                    tar_handle.add(os.path.join(root, file))
-        tar_handle.close()
-
-    def cleanup(self):
-        """
-        Zips the aggregate results, and deletes files for cleanup
-
-        """
-        self.tardir('results', f'{self.output_name}.tar.gz')
-        shutil.rmtree('results')
-        # shutil.rmtree('/opt/genepatt/extracted')
-
-
-    def json_modifications(self):
-        """
-        Modify outputs of the final run.json
-
-        Modifications include:
-        1. correcting string of list
-        2. abslute pathing to relative pathing
-        """
-        with open('results/run.json') as json_file:
-            dict = json.load(json_file)
-        json_file.close()
-
-        for sample in dict['runs'].keys():
-            ## String of list to list
-            for sample_num in range(len(dict['runs'][sample])):
-                try:
-                    sample_name = dict['runs'][sample][sample_num]['Sample name']
-
-                    locations = dict['runs'][sample][sample_num]['Location']
-                    if "|" in locations:
-                        dict['runs'][sample][sample_num]['Location'] = locations.split("|")
-                    else:
-                        dict['runs'][sample][sample_num]['Location'] = locations.strip('][').split(',')
-
-
-                    ## feature bed location relative path conversion
-                    feature_bed_loc = dict['runs'][sample][sample_num]['Feature BED file']
-                    # print(feature_bed_loc)
-                    basename = os.path.basename(feature_bed_loc)
-                    dict['runs'][sample][sample_num]['Feature BED file'] = f"AA_outputs/{sample_name}/{sample_name}_classification/{sample_name}_classification_bed_files/{basename}"
-
-                    ## aa png file
-                    aa_png_location = dict['runs'][sample][sample_num]['AA PNG file']
-                    basename = os.path.basename(aa_png_location)
-                    dict['runs'][sample][sample_num]['AA PNG file'] = f"AA_outputs/{sample_name}/{sample_name}_AA_results/{basename}"
-
-
-                    ## aa pdf file
-                    aa_pdf_location = dict['runs'][sample][sample_num]['AA PDF file']
-                    basename = os.path.basename(aa_pdf_location)
-                    dict['runs'][sample][sample_num]['AA PDF file'] = f"AA_outputs/{sample_name}/{sample_name}_AA_results/{basename}"
-
-                    ## cnv bed file
-                    cnv_bed_location = dict['runs'][sample][sample_num]['CNV BED file']
-                    basename = os.path.basename(cnv_bed_location)
-                    dict['runs'][sample][sample_num]['CNV BED file'] = f"AA_outputs/{sample_name}/{sample_name}_classification/files/{basename}"
-
-                    ## run metadata json
-                    metadata_json_location = dict['runs'][sample][sample_num]['"Run metadata JSON"']
-                    basename = os.path.basename(metadata_json_location)
-                    dict['runs'][sample][sample_num]['AA PDF file'] = f"AA_outputs/{sample_name}/{basename}"
-
-
-
-                except Exception as e:
-                    print(e)
-
-                try:
-                    oncogenes = dict['runs'][sample][sample_num]['Oncogenes']
-                    if "|" in oncogenes:
-                        dict['runs'][sample][sample_num]['Oncogenes'] = oncogenes.split("|")
-                    else:
-                        dict['runs'][sample][sample_num]['Oncogenes'] = oncogenes.strip('][').split(',')
-
-                except:
-                    print(f"Oncogene for sample{sample_num} is None")
-
-
-
-
-        with open('results/run.json', 'w') as json_file:
-            json.dump(dict, json_file)
-        json_file.close()
-
-    def absolute_to_relative_fp(fp, sample_name):
-        """
-        Converts an absolute file path to a relative file path
-        """
-
-        filename = os.path.basename(fp)
-        return f"AA_outputs/{sample_name}/{sample_name}_classification/"
-
-# def run_make_results_table(make_results_dict):
-#     """
-#     Runs Jens's 'make_input.sh' and 'make_results_table.py'
-#     on extracted AA_results
-#     """
-#     for sample in make_results_dict.keys():
-#         name = make_results_dict[sample]['name']
-#         input = make_results_dict[sample]['ac_input']
-#         profiles = make_results_dict[sample]['profiles_tsv']
-#         cnv = make_results_dict[sample]['cnv_calls']
-#
-#
-#         ## run make_results_table.py for each sample
-#         make_results_script = f'python3 /opt/genepatt/make_results_table.py --input {input} --classification_file {profiles}'
-#         print(make_results_script)
-#         os.system(make_results_script)
-
+from AmpliconSuiteAggregatorFunctions import *
 
 
 
 if __name__ == "__main__":
-    # os.environ["AA_DATA_REPO"] = os.path.join('/opt/genepatt', '.data_repo')
 
     parser = argparse.ArgumentParser()
     parser.add_argument("-flist", "--filelist", type = str,
@@ -374,9 +11,6 @@ if __name__ == "__main__":
     parser.add_argument("--output_name", type = str, help = "Output Prefix", default = "aggregated")
 
     args = parser.parse_args()
-
-
-    # os.system(f'bash download_ref.sh {REFERENCE}')
 
 
     filelist = args.filelist
